# feat: optimize performance of repository::random()