@nikophil 2025-04-09
nice!

I'm gonna keep this one opened to test the github action I think