@nikophil 2025-04-13
I'm merging this PR, because I cannot have the "random" bench being usable: it has too much versatility, so we can"t rely on it (if anyone can help for this that would be awesome)